preparing to postpone hacking javaparing
It is insaly dificult task. JavaParing project depends to much of the
awt object and it actualy relies heavely on the state of the awt values.
Even the paring algoritm it looks for the data in this objest and then
initializes the paring. I will probaly abandon this part of the quest
that involvs JavaParing project anleas the creator of the project is not
helping out with the decupaling task.